third_party/WebKit/LayoutTests/accessibility/aria-labelledby-overrides-label-expected.txt

Label
A
B
C


This tests that if aria-labelledby is used, then label elements are not used

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS axTarget.name is 'A B C'
PASS axTarget.nameElementCount() is 3
PASS successfullyParsed is true

TEST COMPLETE
